src/requirements/dev.txt
# Server
django==5.0.4
djangorestframework==3.15.1
drf_spectacular==0.27.2

# Tests
pytest==8.1.1
requests==2.31.0

# Type checking
mypy==1.9.0
django-stubs==4.2.7
types_requests==2.31.0.20240406
djangorestframework_stubs==3.14.5
